+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_03_01_082933) do
+ActiveRecord::Schema.define(version: 2021_03_16_102910) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
   enable_extension "plpgsql"
@@ -31,9 +31,6 @@ ActiveRecord::Schema.define(version: 2021_03_01_082933) do
     t.string "aid_type"
     t.string "form_state"
     t.string "level"
-    t.string "accountable_organisation_name"
-    t.string "accountable_organisation_reference"
-    t.string "accountable_organisation_type"
     t.uuid "extending_organisation_id"
     t.string "recipient_country"
     t.string "geography"