+++ db/seeds/organisations.rb
@@ -4,10 +4,11 @@ organisations = YAML.safe_load(File.read(File.join(Rails.root, "db", "seeds", "o
 organisations.each do |organisation|
   organisation_params = {
     name: organisation["name"],
+    beis_organisation_reference: organisation["short_name"],
     organisation_type: organisation["type"],
     language_code: "en",
     default_currency: "gbp",
     service_owner: organisation["service_owner"],
   }
-  Organisation.find_or_create_by(iati_reference: organisation["reference"]).update(organisation_params)
+  Organisation.find_or_create_by(iati_reference: organisation["reference"]).update!(organisation_params)
 end