+++ spec/features/staff/users_can_mark_a_report_as_awaiting_changes_spec.rb
@@ -1,20 +1,29 @@
 RSpec.feature "Users can move reports into awaiting changes & view reports awaiting changes" do
   context "signed in as a BEIS user" do
     let(:beis_user) { create(:beis_user) }
+    let(:organisation) { create(:organisation, users: create_list(:delivery_partner_user, 3)) }
 
     before do
       authenticate!(user: beis_user)
     end
 
     scenario "they can mark a report as awaiting changes" do
-      report = create(:report, state: :in_review)
+      report = create(:report, state: :in_review, organisation: organisation)
 
-      visit report_path(report)
-      click_link t("action.report.request_changes.button")
-      click_button t("action.report.request_changes.confirm.button")
+      perform_enqueued_jobs do
+        visit report_path(report)
+        click_link t("action.report.request_changes.button")
+        click_button t("action.report.request_changes.confirm.button")
+      end
 
       expect(page).to have_content "awaiting changes"
       expect(report.reload.state).to eql "awaiting_changes"
+
+      expect(ActionMailer::Base.deliveries.count).to eq(organisation.users.count)
+
+      organisation.users.each do |user|
+        expect(user).to have_received_email.with_subject(t("mailer.report.awaiting_changes.subject", application_name: t("app.title")))
+      end
     end
 
     context "when the report is already awaiting changes" do